+++ cohttp/src/cohttp.ml
@@ -1,7 +1,6 @@
 module Accept = Accept
 module Auth = Auth
 module Body = Body
-module Client = Client
 module Conf = Conf
 module Connection = Connection [@@deprecated "Connection.t values are useless."]
 module Code = Code
@@ -11,9 +10,13 @@ module Link = Link
 module Request = Request
 module Response = Response
 module S = S
-module Server = Server
 module Path = Path
 module Transfer = Transfer
+
+module Generic = struct
+  module Client = Client
+  module Server = Server
+end
 
 module Private = struct
   module Transfer_io = Transfer_io